make changes to file path to fix index.html not importing script. Added cartsubscript function.
