+++ src/cart/index.html
@@ -9,8 +9,8 @@
     <title>Sleep Outside | Cart</title>
 
     <link rel="stylesheet" href="/css/style.css" />
-    <script src="../js/main.js" type="module"></script>
     <script src="../js/cart.js" type="module"></script>
+    <script src="../js/cartsubscript.js" type="module"></script>
   </head>
 
   <body>
