+++ src/index.html
@@ -5,9 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
-    <!--<script src="../js/main.js" type="module"></script>
-  <script src="../js/product.js" type="module"></script>-->
-    <!--<script src="../js/home.js" type="module"></script>-->
+    <script src="js/cartsubscript.js" type="module"></script>
   </head>
   <body>
     <header class="divider">
