--- src/js/main.js
@@ -1,3 +1,0 @@
-import { setCartSub } from "./cartsubscript.mjs";
-
-window.addEventListener("load", setCartSub);
